Tiltfile: enable live_reload by default
As Tilt is mainly used for development and live testing, out of the box
configuration should provide best possible DX.

In case production Dockerfile needs to be tested, it should be turned
off.

If https://github.com/tilt-dev/tilt/issues/3070, perhaps this could be
improved, so the same Dockerfile can be used in all cases.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>