Adding EnvVarsWithSelectors to disable KSM (#64)
* feat(envVar): added ConfigWithSelectors

Config withSelecotrs incorporetes resourcesWithSelectors adding
extraEnvVar. This is useful to change the envVars injected depending on the labels
of the pod. The initial use-case is to disable the KSM discovery when
injected in non-KSM pods

* fix(extraEnvVars)

removing the extraEnvVars without selectors. They were overlapping and
it was not 100% clear what was and was not considered when computing the
hash.