Makefile: fix condition check in image target
GOOS := ... line does not set shell variable, but make variable, so we
must usue it as $(GOOS), not as $$GOOS.

Also, GitHub Actions use Ubuntu 20.04 which has old version of ash,
which has no '[[' defined.

Finally, compared string in ash must be in quotes, otherwise comparison
does not work as expected.

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>